EZP-24375: rename provider to resolver
