+++ eZ/Publish/Core/settings/search_engines/solr.yml
@@ -9,18 +9,18 @@ parameters:
     ezpublish.search.solr.connection.server: http://localhost:8983/solr/core0
     ezpublish.spi.search.solr.class: eZ\Publish\Core\Search\Solr\Handler
     ezpublish.search.solr.content.gateway.native.class: eZ\Publish\Core\Search\Solr\Content\Gateway\Native
-    ezpublish.search.solr.content.gateway.endpoint_provider.translation_endpoint_provider.class: eZ\Publish\Core\Search\Solr\Content\Gateway\EndpointProvider\TranslationEndpointProvider
+    ezpublish.search.solr.content.gateway.endpoint_resolver.translation_endpoint_resolver.class: eZ\Publish\Core\Search\Solr\Content\Gateway\EndpointResolver\TranslationEndpointResolver
     ezpublish.spi.search.solr.content_handler.class: eZ\Publish\Core\Search\Solr\Content\Handler
     ezpublish.search.solr.content.document_mapper.translation_document_mapper.class: eZ\Publish\Core\Search\Solr\Content\DocumentMapper\TranslationDocumentMapper
     ezpublish.search.solr.result_extractor.loading.class: eZ\Publish\Core\Search\Solr\Content\ResultExtractor\LoadingResultExtractor
     ezpublish.spi.search.solr.location_handler.class: eZ\Publish\Core\Search\Solr\Content\Location\Handler
 
 services:
-    ezpublish.search.solr.content.gateway.endpoint_provider.translation_endpoint_provider:
-        class: %ezpublish.search.solr.content.gateway.endpoint_provider.translation_endpoint_provider.class%
+    ezpublish.search.solr.content.gateway.endpoint_resolver.translation_endpoint_resolver:
+        class: %ezpublish.search.solr.content.gateway.endpoint_resolver.translation_endpoint_resolver.class%
 
-    ezpublish.search.solr.content.gateway.endpoint_provider:
-        alias: ezpublish.search.solr.content.gateway.endpoint_provider.translation_endpoint_provider
+    ezpublish.search.solr.content.gateway.endpoint_resolver:
+        alias: ezpublish.search.solr.content.gateway.endpoint_resolver.translation_endpoint_resolver
 
     ezpublish.search.solr.content.document_mapper.translation_document_mapper:
         class: %ezpublish.search.solr.content.document_mapper.translation_document_mapper.class%
@@ -50,7 +50,7 @@ services:
         class: %ezpublish.search.solr.content.gateway.native.class%
         arguments:
             - @ezpublish.search.solr.content.gateway.client.http.stream
-            - @ezpublish.search.solr.content.gateway.endpoint_provider
+            - @ezpublish.search.solr.content.gateway.endpoint_resolver
             - @ezpublish.search.solr.content.gateway.endpoint_registry
             - @ezpublish.search.solr.content.criterion_visitor.aggregate
             - @ezpublish.search.solr.content.sort_clause_visitor.aggregate
@@ -75,7 +75,7 @@ services:
         class: %ezpublish.search.solr.content.gateway.native.class%
         arguments:
             - @ezpublish.search.solr.content.gateway.client.http.stream
-            - @ezpublish.search.solr.content.gateway.endpoint_provider
+            - @ezpublish.search.solr.content.gateway.endpoint_resolver
             - @ezpublish.search.solr.content.gateway.endpoint_registry
             - @ezpublish.search.solr.location.criterion_visitor.aggregate
             - @ezpublish.search.solr.location.sort_clause_visitor.aggregate
